split the packet processing logic from the controller file
